requirements.txt
opencv-python==4.8.0.74
pytesseract==0.3.6
python-Levenshtein
numpy